requirements.txt
# Core dependencies
numpy>=1.24.3
pandas>=2.0.2
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn>=1.2.2
biopython>=1.81

# Quantum computing
qiskit>=0.43.0
qiskit-aer>=0.12.0  # Qiskit's high-performance simulator
qiskit-ibm-runtime>=0.11.0  # For IBM Quantum access if needed
pennylane>=0.32.0
pennylane-qiskit>=0.32.0  # PennyLane plugin for Qiskit

# PyTorch (install appropriate version for your system)
# For CUDA 11.8 (recommended if you have NVIDIA GPU):
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Alternative PyTorch installation (CPU-only or different CUDA version):
# See https://pytorch.org/get-started/locally/ for the right command
# For CPU-only: pip3 install torch torchvision torchaudio

# Additional utilities
tqdm>=4.65.0  # Progress bars
joblib>=1.2.0  # For parallel processing
